requirements.txt
requests
python-dotenv
pandas
duckdb
pyarrow
streamlit
altair
